backend/pyproject.toml
[project]
name = "autoshop-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.30.0",
    "langgraph>=0.3.0,<0.4.0",
    "anthropic>=0.40.0,<1.0.0",
    "deepgram-sdk>=3.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "redis[hiredis]>=5.0.0",
    "boto3>=1.34.0",
    "aioboto3>=13.0.0,<16.0.0",
    "twilio>=9.0.0",
    "sendgrid>=6.11.0",
    "PyJWT>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=3.2.0,<4.0.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "pydantic-settings>=2.0.0",
    "reportlab>=4.2.0",
    "pillow>=10.0.0",
    "qdrant-client[async]>=1.9.0,<2.0.0",
    "openai>=1.30.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "pytest-cov>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
